+++ src/index.js
@@ -14,24 +14,83 @@
 
 import { GoogleLogger } from './google-logger';
 import { CoralogixLogger } from './coralogix-logger';
+import { CoralogixErrorLogger } from './coralogix-error-logger';
+import { respondRobots } from './robots.js';
+import { respondUnpkg } from './unpkg.js';
 
-// const GoogleLogger = require('./google-logger');
-// const CoralogixLogger = require('./coralogix-logger');
+function respondError(message, status, e, req) {
+  const headers = new Headers();
+  const msg = e && e.message ? `${message}: ${e.message}` : message;
+
+  headers.set('Content-Type', 'text/plain; charset=utf-8');
+  headers.set('X-Error', msg);
+
+  const response = new Response(`${msg}\n`, {
+    status,
+    headers,
+  });
+  console.error(msg);
+  try {
+    const c = new CoralogixErrorLogger(req);
+    c.logError(status, message);
+  } catch (err) {
+    console.error(`error logging error: ${err.message}`);
+  }
+  return response;
+}
+
+function hashCode(s) {
+  // eslint-disable-next-line no-bitwise
+  s.split('').reduce((a, b) => (((a << 5) - a) + b.charCodeAt(0)) | 0, 0);
+}
 
 async function main(req) {
   try {
-    const body = await req.json();
+    if (req.method === 'GET' && new URL(req.url).pathname.startsWith('/robots.txt')) {
+      return respondRobots(req);
+    }
+    if (req.method === 'GET' && new URL(req.url).pathname.startsWith('/.rum/web-vitals')) {
+      return respondUnpkg(req);
+    }
+    if (req.method === 'GET' && new URL(req.url).pathname.startsWith('/.rum/@adobe/helix-rum')) {
+      return respondUnpkg(req);
+    }
+    const body = req.method === 'GET'
+      ? JSON.parse(new URL(req.url).searchParams.get('data'))
+      : await req.json();
 
     const headers = new Headers();
     headers.set('Content-Type', 'text/plain; charset=utf-8');
 
-    const { weight = 1, id, cwv = {} } = body;
+    const {
+      weight = 1,
+      id = req.method === 'GET' ? `${hashCode(req.url)}-${new Date().getTime()}-${Math.random().toString(16).substr(2, 14)}` : undefined,
+      cwv = {},
+      referer, referrer,
+      generation, checkpoint,
+      target,
+      source,
+    } = body;
 
-    const c = new CoralogixLogger(req);
-    c.logRUM(cwv, id, weight);
+    if (!id) {
+      return respondError('id field is required', 400, undefined, req);
+    }
+    if (!weight || typeof weight !== 'number') {
+      return respondError('weight must be a number', 400, undefined, req);
+    }
+    if (typeof cwv !== 'object') {
+      return respondError('cwv must be an object', 400, undefined, req);
+    }
 
-    const g = new GoogleLogger(req);
-    g.logRUM(cwv, id, weight);
+    try {
+      const c = new CoralogixLogger(req);
+      c.logRUM(cwv, id, weight, referer || referrer, generation, checkpoint, target, source);
+
+      const g = new GoogleLogger(req);
+      g.logRUM(cwv, id, weight, referer || referrer, generation, checkpoint, target, source);
+    } catch (err) {
+      return respondError(`Could not collect RUM: ${err.message}`, 500, err, req);
+    }
 
     const response = new Response('rum collected.', {
       status: 201,
@@ -40,72 +99,17 @@ async function main(req) {
 
     return response;
   } catch (e) {
-    const headers = new Headers();
-
-    console.error(e);
-
-    const message = `RUM collection expects a JSON POST or PUT body: ${e.message}`;
-
-    headers.set('Content-Type', 'text/plain; charset=utf-8');
-    headers.set('X-Error', message);
-
-    const response = new Response(`${message}\n`, {
-      status: 400,
-      headers,
-    });
-    console.error('RUM collection expects a JSON POST or PUT body.');
-
-    return response;
+    return respondError('RUM Collector expects POST body as JSON', 400, e, req);
   }
 }
 
-addEventListener('fetch', async (event) => {
-  // NOTE: By default, console messages are sent to stdout (and stderr for `console.error`).
-  // To send them to a logging endpoint instead, use `console.setEndpoint:
-  // console.setEndpoint("my-logging-endpoint");
-
+async function handler(event) {
   // Get the client reqest from the event
   const req = event.request;
+  return main(req);
+}
 
-  event.respondWith(await main(req, {}));
-
-  /*
-
-  // Make any desired changes to the client request.
-  req.headers.set('Host', 'example.com');
-
-  // We can filter requests that have unexpected methods.
-  const VALID_METHODS = ['GET'];
-  if (!VALID_METHODS.includes(req.method)) {
-    const response = new Response('This method is not allowed', {
-      status: 405,
-    });
-    // Send the response back to the client.
-    event.respondWith(response);
-    return;
-  }
-
-  const { method } = req;
-  const urlParts = req.url.split('//').pop().split('/');
-  const path = `/${urlParts.join('/')}`;
-
-  // If request is a `GET` to the `/` path, send a default response.
-  if (method === 'GET' && path === '/') {
-    const headers = new Headers();
-    headers.set('Content-Type', 'text/plain; charset=utf-8');
-    const response = new Response('Hallo Stefan\n', {
-      status: 200,
-      headers,
-    });
-    // Send the response back to the client.
-    event.respondWith(response);
-    return;
-  }
-
-  // Catch all other requests and return a 404.
-  const response = new Response('The page you requested could not be found', {
-    status: 404,
-  });
-  // Send the response back to the client.
-  event.respondWith(response); */
+// eslint-disable-next-line no-restricted-globals
+addEventListener('fetch', (event) => {
+  event.respondWith(handler(event));
 });